fix: move status filter helper out of features
