+++ source/features/extend-conversation-status-filters.tsx
@@ -5,8 +5,8 @@ import {$} from 'select-dom';
 
 import features from '../feature-manager.js';
 import SearchQuery from '../github-helpers/search-query.js';
+import {setStatusFilter} from '../helpers/set-status-filter.js';
 import observe from '../helpers/selector-observer.js';
-import {setStatusFilter} from './extend-conversation-status-filters-helpers.js';
 
 function addMergeLink(lastLink: HTMLAnchorElement): void {
 	// It's shouldn't be added in issues list
